requirements.txt
streamlit==1.26.0
openai==0.27.0
torch==2.0.0
torchvision==0.15.1
pillow==9.4.0
moviepy==1.0.3
numpy==1.24.2
python-dotenv==1.0.0
